# src/openapi.yaml
openapi: 3.1.0
info:
  title: Time-addressable Media Store
  description: 'The Time-addressable Media Store (TAMS) is used for storing segmented
    media Flows.

    The TAMS consists of one or more Storage Backends for the media Flow Segment Objects
    and a service providing a database index of the Flow Segments.

    This document is a specification of the service API.


    See the [bbc/tams](https://github.com/bbc/tams) repository for more background
    on TAMS.


    **Note**: the examples provided in this specification are for a system which uses
    a Storage Backend that provides HTTP URLs for uploading and downloading Media
    Objects in buckets.

    This could for example be implemented using an AWS S3 compatible store using presigned
    URLs or by a simple file system storage with an HTTP frontend.

    Clients should parse the [/service](#/operations/GET_service) endpoint and handle
    the Storage Backend type in use appropriately.

    '
  version: '8.0'
  contact:
    name: BBC R&D - Cloud-Fit Production Team
    email: cloudfit-opensource@rd.bbc.co.uk
    url: https://github.com/bbc/tams
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
- url: http://localhost:4010
  description: Local mock of API
- url: https://example.com/tams/{version}
  description: Example TAMS service
  variables:
    version:
      description: API version
      default: v8.0
security:
- bearer_token_auth: []
- basic_auth: []
paths:
...
  /flows:
    head:
      summary: List Flows
      description: Return Flows path headers
      operationId: HEAD_flows
      tags:
      - Flows
      parameters:
      - name: source_id
        in: query
        description: Filter on Source identifier.
        schema:
          $ref: vendor/bbc-tams/api/schemas/uuid.json
      - name: timerange
        in: query
        description: Filter on Flows that overlap the given timerange.
        schema:
          default: _
          $ref: vendor/bbc-tams/api/schemas/timerange.json
      - name: include_timerange
        in: query
        x-tamoss-extension: true
        description: TAMOSS extension. Include each listed Flow's computed content
          timerange in the response.
        schema:
          default: false
          type: boolean
      - name: format
        in: query
        description: Filter on Flow format.
        schema:
          $ref: vendor/bbc-tams/api/schemas/content-format.json
      - name: codec
        in: query
        description: Filter on Flow codec.
        schema:
          $ref: vendor/bbc-tams/api/schemas/mime-type.json
      - name: label
        in: query
        description: Filter on Flows that have the given label.
        schema:
          type: string
      - name: tag.{name}
        in: query
        description: 'Filter on flows that have a tag named {name} with a value in
          the given comma-separated list of values.

          The {name} and the value MUST be URL encoded where special characters are
          present.

          Where the tag''s value is a string, at least one of the given values will
          match.

          Where the tag''s value is an array, at least one value in the array will
          match at least one of the given values.

          Partial string matches of the values are not valid.

          '
        schema:
          $ref: vendor/bbc-tams/api/schemas/url-tag-list.json
      - name: tag_exists.{name}
        in: query
        description: 'Filter on Flows that have a tag named {name} regardless of value.

          {name} MUST be URL encoded where special characters are present.

          If set to true then the presence of the tag is filtered for.

          If set to false then its absence is.

          If left out then no filtering on tag presence is performed.

          '
        schema:
          type: boolean
      - name: frame_width
        in: query
        description: Filter on video Flows that have the given frame width.
        schema:
          type: integer
      - name: frame_height
        in: query
        description: Filter on video Flows that have the given frame height.
        schema:
          type: integer
      - $ref: '#/components/parameters/trait_resource_paged_key'
      - $ref: '#/components/parameters/trait_paged_limit'
      responses:
        '200':
          description: ''
          headers:
            Link:
              description: Provides references to cursors for paging. Only the 'rel'
                attribute with value 'next' and a link to the next page is currently
                supported. If 'next' is not present then it is the last page.
              schema:
                type: string
            X-Paging-Limit:
              description: Identifies the current limit being used for paging. This
                may not match the requested value if the requested value was too high
                for the service implementation
              schema:
                type: integer
            X-Paging-NextKey:
              description: Opaque string that can be supplied to the `page` query
                parameter to get the next page of results.
              schema:
                type: string
          content:
            application/json:
              schema:
                type: string
        '400':
          description: Bad request. Invalid query options.
    get:
      summary: List Flows
      description: List the Flows registered in the TAMS service instance.
      operationId: GET_flows
      tags:
      - Flows
      parameters:
      - name: source_id
        in: query
        description: Filter on Source identifier.
        schema:
          $ref: vendor/bbc-tams/api/schemas/uuid.json
      - name: timerange
        in: query
        description: Filter on Flows that overlap the given timerange. An empty timerange
          returns Flows with no content.
        schema:
          default: _
          $ref: vendor/bbc-tams/api/schemas/timerange.json
      - name: include_timerange
        in: query
        x-tamoss-extension: true
        description: TAMOSS extension. Include each listed Flow's computed content
          timerange in the response.
        schema:
          default: false
          type: boolean
      - name: format
        in: query
        description: Filter on Flow format.
        schema:
          $ref: vendor/bbc-tams/api/schemas/content-format.json
      - name: codec
        in: query
        description: Filter on Flow codec.
        schema:
          $ref: vendor/bbc-tams/api/schemas/mime-type.json
      - name: label
        in: query
        description: Filter on Flows that have the given label.
        schema:
          type: string
      - name: tag.{name}
        in: query
        description: 'Filter on flows that have a tag named {name} with a value in
          the given comma-separated list of values.

          The {name} and the value MUST be URL encoded where special characters are
          present.

          Where the tag''s value is a string, at least one of the given values will
          match.

          Where the tag''s value is an array, at least one value in the array will
          match at least one of the given values.

          Partial string matches of the values are not valid.

          '
        schema:
          $ref: vendor/bbc-tams/api/schemas/url-tag-list.json
      - name: tag_exists.{name}
        in: query
        description: 'Filter on Flows that have a tag named {name} regardless of value.

          {name} MUST be URL encoded where special characters are present.

          If set to true then the presence of the tag is filtered for.

          If set to false then its absence is.

          If left out then no filtering on tag presence is performed.

          '
        schema:
          type: boolean
      - name: frame_width
        in: query
        description: Filter on video Flows that have the given frame width.
        schema:
          type: integer
      - name: frame_height
        in: query
        description: Filter on video Flows that have the given frame height.
        schema:
          type: integer
      - $ref: '#/components/parameters/trait_resource_paged_key'
      - $ref: '#/components/parameters/trait_paged_limit'
      responses:
        '200':
          description: ''
          headers:
            Link:
              description: Provides references to cursors for paging. Only the 'rel'
                attribute with value 'next' and a link to the next page is currently
                supported. If 'next' is not present then it is the last page.
              schema:
                type: string
            X-Paging-Limit:
              description: Identifies the current limit being used for paging. This
                may not match the requested value if the requested value was too high
                for the service implementation
              schema:
                type: integer
            X-Paging-NextKey:
              description: Opaque string that can be supplied to the `page` query
                parameter to get the next page of results.
              schema:
                type: string
          content:
            application/json:
              example:
                $ref: vendor/bbc-tams/api/examples/flows-get-200.json
              schema:
                type: array
                items:
                  $ref: vendor/bbc-tams/api/schemas/flow.json
        '400':
          description: Bad request. Invalid query options.
...
  /flows/{flowId}/tags:
    parameters:
    - name: flowId
      in: path
      required: true
      schema:
        $ref: vendor/bbc-tams/api/schemas/uuid.json
      description: The Flow identifier.
    head:
      summary: List Flow Tags
      description: Return Flow tags path headers
      operationId: HEAD_flows-flowId-tags
      tags:
      - Flows
      responses:
        '200':
          $ref: '#/components/responses/trait_resource_listing_head_200'
        '404':
          description: The requested Flow does not exist.
    get:
      summary: List Flow Tags
      description: Returns the Flow tags.
      operationId: GET_flows-flowId-tags
      tags:
      - Flows
      responses:
        '200':
          description: ''
          content:
            application/json:
              example:
                $ref: vendor/bbc-tams/api/examples/flow-tags-get-200.json
              schema:
                $ref: vendor/bbc-tams/api/schemas/tags.json
        '404':
          description: The requested Flow does not exist.
  /flows/{flowId}/tags/{name}:
    parameters:
    - name: name
      in: path
      required: true
      schema:
        type: string
      description: The tag name. {name} MUST be URL encoded where special characters
        are present.
    - name: flowId
      in: path
      required: true
      schema:
        $ref: vendor/bbc-tams/api/schemas/uuid.json
      description: The Flow identifier.
    head:
      summary: Flow Tag Value
      description: Return Flow tag path headers
      operationId: HEAD_flows-flowId-tags-name
      tags:
      - Flows
      responses:
        '200':
          $ref: '#/components/responses/trait_resource_info_head_200'
        '404':
          description: The requested Flow or tag does not exist.
    get:
      summary: Flow Tag Value
      description: Return the tag value associated with the tag name.
      operationId: GET_flows-flowId-tags-name
      tags:
      - Flows
      responses:
        '200':
          description: ''
          content:
            application/json:
              example: '"full"

                '
              schema:
                type: string
        '404':
          description: The requested Flow or tag does not exist.
    put:
      summary: Create or Update Flow Tag
      description: Create or update the tag.
      operationId: PUT_flows-flowId-tags-name
      tags:
      - Flows
      requestBody:
        content:
          application/json:
            example: '"proxy"

              '
            schema:
              type: string
        required: true
      responses:
        '204':
          description: No content. The tag has been created or updated.
        '400':
          description: Bad request. Invalid Flow tag value.
        '403':
          description: Forbidden. You do not have permission to modify this Flow.
            It may be marked read-only.
        '404':
          description: The requested Flow does not exist.
    delete:
      summary: Delete Flow Tag
      description: Delete the tag.
      operationId: DELETE_flows-flowId-tags-name
      tags:
      - Flows
      responses:
        '204':
          description: No content. The Flow tag has been deleted.
        '403':
          description: Forbidden. You do not have permission to modify this Flow.
            It may be marked read-only.
        '404':
          description: The requested Flow ID in the path is invalid.
  /flows/{flowId}/description:
    parameters:
    - name: flowId
      in: path
      required: true
      schema:
        $ref: vendor/bbc-tams/api/schemas/uuid.json
      description: The Flow identifier.
    head:
      summary: Flow Description
      description: Return Flow description path headers
      operationId: HEAD_flows-flowId-description
      tags:
      - Flows
      responses:
        '200':
          $ref: '#/components/responses/trait_resource_info_head_200'
        '404':
          $ref: '#/components/responses/trait_resource_info_head_404'
    get:
      summary: Flow Description
      description: Returns the Flow description property. This should be a human-readable
        description that may be showed in detailed views of Flows. The description
        should be longer and more detailed than `label`.
      operationId: GET_flows-flowId-description
      tags:
      - Flows
      responses:
        '200':
          description: ''
          content:
            application/json:
              example: '"video capture"

                '
              schema:
                type: string
        '404':
          description: The requested Flow does not exist.
    put:
      summary: Create or Update Flow Description
      description: Create or update the description property. This should be a human-readable
        description that may be showed in detailed views of Flows. The description
        should be longer and more detailed than `label`.
      operationId: PUT_flows-flowId-description
      tags:
      - Flows
      requestBody:
        content:
          application/json:
            example: '"Big Buck Bunny video-only capture"

              '
            schema:
              type: string
        required: true
      responses:
        '204':
          description: No content. The description has been created or updated.
        '400':
          description: Bad request. Invalid Flow description.
        '403':
          description: Forbidden. You do not have permission to modify this Flow.
            It may be marked read-only.
        '404':
          description: The requested Flow does not exist.
    delete:
      summary: Delete Flow Description
      description: Delete the description property.
      operationId: DELETE_flows-flowId-description
      tags:
      - Flows
      responses:
        '204':
          description: No content. The Flow description property has been deleted.
        '403':
          description: Forbidden. You do not have permission to modify this Flow.
            It may be marked read-only.
        '404':
          description: The requested Flow ID in the path is invalid.
...
components:
  responses:
    trait_resource_listing_head_200:
      description: ''
      content:
        application/json:
          schema:
            type: string
    trait_resource_info_head_200:
      description: ''
      content:
        application/json:
          schema:
            type: string
...
    trait_resource_info_head_404:
      description: Resource was not found.
  parameters:
    trait_paged_limit:
      name: limit
      in: query
      description: Restrict the response to the specified number of results. Service
        implementations may specify their own default and maximum for the limit
      schema:
        type: integer
    trait_resource_paged_key:
      name: page
      in: query
      description: Opaque string used by backend to access a specific page of results.
        Clients should read the next URL from the `Link` header returned with responses,
        or use value of the returned X-Paging-NextKey header. If not supplied, the
        first page is accessed. Service implementations should ensure a consistent
        sort order is applied to pages of results.
      schema:
        type: string
  securitySchemes:
    basic_auth:
      type: http
      scheme: basic
      description: 'HTTP basic authentication using a username and password.


        This option should only be used in constrained cases where Bearer tokens are
        impractical, since the password is intrinsically long-lived, and managing
        user credentials across systems is difficult compared to using centralised
        auth (e.g. via OAuth2) with Bearer tokens

        '
    bearer_token_auth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 'HTTP bearer token authentication.


        The method through which a client should acquire a bearer token is assumed
        to be an OAuth2 grant (either the Authorization Code or Client Credentials
        grant), however the OpenAPI `OAuth Flow` object cannot be used because it
        requires that the relevant URLs are known in advance.


        Clients should provide options to allow one (or both) grant types to be configured,
        as appropriate.


        Where tokens are not JWTs, or are acquired through some other mechanism, the
        API key method should be used instead.

        '
...